+++ src/ParseMapfile.hpp
@@ -4,8 +4,8 @@
 #ifndef _READMAPFILE_H_
 #define _READMAPFILE_H_
 
-class SeqStatsGenome;
+class SeqStatsGenomeSSP;
 
-void read_mapfile(SeqStatsGenome &genome);
+void read_mapfile(SeqStatsGenomeSSP &genome);
 
 #endif /* _READMAPFILE_H_ */